Move camera window to its own module
